requirements.txt
mistralai
pillow
gradio